changes to index.php and css
I had to redo the table and put in th’s for headers instead of all td’s.
I styled the table to at least get it looking somewhat functional with
the help from one of my books.
I have no idea that the rogue false is doing on the displayed page.